fix(misc): move cross-cutting req-context accessors to config-free req-context.ts
permissions.ts (config-coupled) hosting the resource/resourceType/
bypassPermissions/publicOperation accessors made query-advice.ts — a
config-free unit-tested util — transitively load and validate #config at
import, breaking its unit spec under the config-less 'unit' project.

Move the accessor definitions into the config-free req-context.ts; permissions.ts
re-exports them as a facade so permissions.<accessor> namespace consumers keep
working; config-free consumers import directly from req-context.ts.

405 unit tests pass (was failing); ratchet unchanged (1354); behavior identical.
