+++ api/src/catalog/router.ts
@@ -6,7 +6,7 @@ import mongo from '#mongo'
 import { findDatasets } from '../datasets/service.js'
 import * as apiKeyUtils from '../misc/utils/api-key.ts'
 import * as cacheHeaders from '../misc/utils/cache-headers.ts'
-import { setReqResourceType } from '../misc/utils/permissions.ts'
+import { setReqResourceType } from '../misc/utils/req-context.ts'
 import { reqPublicationSite, reqMainPublicationSite, setReqPublicationSite } from '../misc/utils/publication-sites.ts'
 import { buildDcatCatalog } from './operations.ts'
 import { findCatalogDatasets, getCatalogApiDocs } from './service.ts'
