+++ api/src/misc/utils/capture.ts
@@ -5,7 +5,7 @@ import pump from './pipe.ts'
 import * as rateLimiting from './rate-limiting.ts'
 import debug from 'debug'
 import * as permissionsUtils from './permissions.ts'
-import { reqResource } from './permissions.ts'
+import { reqResource } from './req-context.ts'
 import { reqPublicBaseUrl } from './public-base-url.ts'
 import resolvePath from 'resolve-path'
 import { internalError } from '@data-fair/lib-node/observer.js'
