+++ api/src/misc/utils/req-context.ts
@@ -2,13 +2,19 @@
 // Pattern follows @data-fair/lib-express session.ts / site.ts: symbol keys,
 // typed accessors, casts contained in the factory, get() throws when the
 // middleware that sets the value was not applied.
-// Accessors are defined topically, next to the code that owns the context
-// (e.g. reqSettingsParams in settings/middlewares.ts) — see the placement
-// table in docs/architecture/code-conventions.md §2. This module only hosts
-// the factory and req-free context helpers.
+// This module hosts the factory, the req-free context helpers, AND the
+// cross-cutting request-context accessors (resource / resourceType /
+// bypassPermissions / publicOperation) consumed across modules. Those live here
+// — not in their semantic-owner domain module (permissions.ts) — because this
+// module is config-free: a config-free consumer (e.g. the unit-tested pure util
+// query-advice.ts) must be able to import an accessor without transitively
+// pulling in `#config` validation. Module-LOCAL accessors still live next to
+// the code that owns them (e.g. reqSettingsParams in settings/middlewares.ts).
+// See docs/architecture/code-conventions.md §2.
 import type { IncomingMessage } from 'node:http'
 import type { Request } from 'express'
 import { reqSession, type User, type Account } from '@data-fair/lib-express'
+import type { Resource, ResourceType, BypassPermissions } from '#types'
 
 export type ReqContext<T> = {
   set: (req: IncomingMessage, value: T) => void
@@ -35,6 +41,29 @@ export const defineReqContext = <T>(name: string, legacyProp?: string): ReqConte
     getOptional: (req) => (req as any)[key] ?? (legacyProp ? (req as any)[legacyProp] : undefined)
   }
 }
+
+// ---- cross-cutting request-context accessors ----
+// resource / resourceType are the requested resource and its kind; bypassPermissions
+// is set by api-key / application-key middlewares; publicOperation is set by the
+// permissions middleware and read by cache-headers / query-advice / thumbnails.
+// legacyProp keeps the legacy `req.<prop>` reads/writes working while datasets
+// (Phase 6) and api-compat (Phase 7) still mutate them directly.
+const resourceCtx = defineReqContext<Resource>('resource', 'resource')
+export const setReqResource = resourceCtx.set
+export const reqResource = resourceCtx.get
+export const reqResourceOptional = resourceCtx.getOptional
+
+const resourceTypeCtx = defineReqContext<ResourceType>('resourceType', 'resourceType')
+export const setReqResourceType = resourceTypeCtx.set
+export const reqResourceType = resourceTypeCtx.get
+
+const bypassPermissionsCtx = defineReqContext<BypassPermissions>('bypassPermissions', 'bypassPermissions')
+export const setReqBypassPermissions = bypassPermissionsCtx.set
+export const reqBypassPermissions = bypassPermissionsCtx.getOptional
+
+const publicOperationCtx = defineReqContext<boolean>('publicOperation', 'publicOperation')
+export const setReqPublicOperation = publicOperationCtx.set
+export const reqPublicOperation = publicOperationCtx.getOptional
 
 // ---- audit-log context without req (same EventLog output as passing { req }) ----
 export type LogContext = { user?: User, account?: Account, ip?: string, host?: string }
