+++ docs/plans/2026-06-15-refactor-misc-utils.md
@@ -332,7 +332,7 @@ Branch off `master` **after 5a merges** (avoids churn collisions on shared impor
 - Any `req.resource` read that is unconditional today but should tolerate absence (or vice-versa) — preserve the exact `reqResource` vs `reqResourceOptional` choice that matches current runtime behavior; if the current behavior looks buggy, **park it**, do not switch.
 - `permissions.ts:363` `await req.resource` on a non-promise — preserved as `await reqResource(req)` (no-op await). If confirmed pointless, park a cleanup.
 - `api-key.ts` `_readApiKey` typing gap — ALREADY RESOLVED: `Resource` carries `& { _readApiKey?: { current, previous } }` in `api/types/index.ts`. No action needed in A4.
-- **(parked, found in A5)** `query-advice.ts` now imports `permissions.ts` (for `reqPublicOperation`), which eagerly validates `#config`/`#mongo` at module load. Its unit test (`query-advice.unit.spec.ts`) still passes under the standard `NODE_CONFIG_DIR` harness, but the module can no longer load in a bare Node context. Do NOT "fix" by relocating accessors out of their topical home (`permissions.ts`) — that violates the placement-table convention. The convention-respecting fix, if ever pursued, is to make `permissions.ts` config validation lazy (not at import time). Low priority; the test passes. Related: [[feedback_unit_tests_pure_functions]].
+- **(RESOLVED 2026-06-16)** `query-advice.ts` importing `permissions.ts` (config-coupled) broke its unit spec under the `unit` project (no config dir → `config.ts` import-time `assertValid` throws). Fix: moved the cross-cutting accessors (`resource`/`resourceType`/`bypassPermissions`/`publicOperation`) from `permissions.ts` into the config-free `req-context.ts`; `permissions.ts` re-exports them as a facade; config-free consumers (query-advice et al.) import directly from `req-context.ts`. 405 unit tests pass; ratchet unchanged (1354). Convention §2 updated with the config-free-home rule. (The earlier note here was wrong — relocating accessors to a config-free module IS the right fix, not lazy config validation.)
 - **(parked, found in A2)** `cache-headers.ts:39` `new Date(resource[dateKey] || resource.updatedAt)` is `string | undefined` because `Dataset.updatedAt`/`finalizedAt` are optional in the generated types — `new Date(undefined)` has no overload (TS2769). Runtime is fine (datasets always have updatedAt). Not suppressible without a `!`/cast (forbidden). Park: fix by making `updatedAt` required in the dataset schema, or guard. Also parked: the 4 `req.query` ParsedQs reads in cache-headers handlers (lines 65/84/86) — directly-mounted middleware MUST keep express `RequestHandler` typing (the sibling `permissions.ts` project-`Request` typing only works because its call sites cast `as RequestHandler`, which is Phase-6 router scope here). Revisit when the cache-headers handlers can be re-exported through a `RequestHandler`-cast factory during Phase 6.
 
 ---
